backend/requirements.txt
opencv-python
pytesseract
pillow
google-generativeai
numpy
flask
flask-cors
sqlalchemy
python-dotenv
requests
